Refactor/extract client-related interfaces and functions to common/client-utils.ts, allowing them to be used in a client, in addition to the tests.
