+++ .prettierignore
@@ -1,0 +1,5 @@
+node_modules
+coverage
+dist
+.idea
+.goose
